refactor(types): clear the long-tail (count guards, configs, connectors)
Guard Builder::delete() results (typed mixed by Larastan) before interpolating
the count in flash messages across the index/settings delete actions; coerce
env()/config() reads in AppServiceProvider and the Saloon connector auth headers
to string; add list<string> to the TMDB request append-to-response params;
BoardGame/Album status & ownership @property; typed credentials for LoginForm;
guarded queue-position reads; and dropped the dead $rating === '0' branch in
the import services (empty() already covers '0').

All 91 tests pass. Baseline: 50 -> 20 errors, zero additions.

+++ app/Livewire/Movies/MovieTmdbSearch.php
@@ -378,7 +378,7 @@ class MovieTmdbSearch extends Component
             return;
         }
 
-        $userId = Auth::id();
+        $userId = (int) Auth::id();
         $showName = $this->title;
         $posterUrl = $this->poster_url ?: null;
         $now = now();
